pyproject.toml
[tool.poetry]
name = "newsletter-app"
version = "0.1.0"
description = ""
authors = ["sunko-seoul <sunko@student.42seoul.kr>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
py-ulid = "^1.0.3"
dependency-injector = "^4.44.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^2.0.36"
mysqlclient = "^2.2.6"
alembic = "^1.14.0"
paramiko = "^3.5.0"
sshtunnel = "^0.4.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.7.1"
jose = "^1.0.0"
jwt = "^1.3.1"
python-jose = "^3.3.0"
python-multipart = "^0.0.20"
bcrypt = "3.2.2"
aiomysql = "^0.2.0"
greenlet = "^3.1.1"
pytest = "^8.3.4"
pytest-mock = "^3.14.0"
freezegun = "^1.5.1"
httpx = "^0.28.1"
pytest-asyncio = "^0.25.1"
pycryptodome = "^3.21.0"
pydantic = {extras = ["email"], version = "^2.10.4"}
langchain = "^0.3.13"
langchain-community = "^0.3.13"
langgraph = "^0.2.60"
markdownify = "^0.14.1"
pre-commit = "^4.0.1"
ipython = "^8.31.0"
tavily-python = "^0.5.0"
langchain-openai = "^0.2.14"
celery = {extras = ["redis"], version = "^5.4.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["./backend"]


[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # The following are specific to Black, you probably don't want those.
  tests/data/
  | profiling/
  | scripts/generate_schema.py  # Uses match syntax
)
'''
# We use the unstable style for formatting Black itself. If you
# want bug-free formatting, you should keep this off. If you want
# stable formatting across releases, you should also keep `preview = true`
# (which is implied by this flag) off.
unstable = true

# Build system information and other project-specific configuration below.
# NOTE: You don't need this in your own Black configuration.
